+++ src/collective/linkcheck/upgrades.py
@@ -1,10 +1,12 @@
-from Products.CMFCore.utils import getToolByName
-from BTrees.IIBTree import IISet
-
-from zope.component import getUtility
-from plone.registry.interfaces import IRegistry
-
 from .interfaces import ISettings
+from BTrees.IIBTree import IISet
+from Products.CMFCore.utils import getToolByName
+from logging import getLogger
+from plone.registry.interfaces import IRegistry
+from zope.component import getUtility
+
+PROFILE_ID = 'profile-collective.linkcheck:default'
+logger = getLogger("collective.linkcheck")
 
 
 def upgrade_tool(tool):
@@ -39,3 +41,14 @@ def upgrade_tool(tool):
 
     registry = getUtility(IRegistry)
     registry.registerInterface(ISettings)
+
+
+def update_registry(context):
+    setup_tool = getToolByName(context, 'portal_setup')
+    setup_tool.runImportStepFromProfile(PROFILE_ID, 'plone.app.registry')
+
+    registry = getUtility(IRegistry)
+    settings = registry.forInterface(ISettings, check=False)
+    if not hasattr(settings, "report_urls_count"):
+        settings.report_urls_count = 20
+    logger.info("Updated registry entries")